pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "micromech"
version = "0.0.1"
description = "Lightweight OLAS mech runtime"
requires-python = ">=3.12"
license = "Apache-2.0"
authors = [
    { name = "David Vilela", email = "david.vilela@gmail.com" },
]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Programming Language :: Python :: 3",
]
dependencies = [
    "pydantic>=2.0.0",
    "peewee>=3.17.0",
    "loguru>=0.7.0",
    "pyyaml>=6.0.3",
]

[project.optional-dependencies]
chain = [
    "iwa>=0.5.5",
]
llm = [
    "llama-cpp-python>=0.2.0",
    "huggingface-hub>=0.19.0",
]
web = [
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "jinja2>=3.1.0",
]
cli = [
    "typer>=0.9.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
micromech = "micromech.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/micromech"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
